int -> float

+++ src/JirungEngine/ObjectModule/Physics/Physics.hpp
@@ -7,8 +7,8 @@
 
 namespace JirungEngine {
     class Physics {
-        public: unsigned short int elasticity;  // max: 1   if x > 1   ->   x = 1
-        public: unsigned short int density;     // max: 1
+        public: float elasticity;  // max: 1   if x > 1   ->   x = 1
+        public: float density;     // max: 1
         public: Vector velocity;
         public: Vector acceleration;
         public: Vector gravity;
